apps/suap_perfil/requirements-app.txt
#-e ../ege_django_theme
ege-django-theme==3.7

python-social-auth==0.3.6
social-auth-app-django==2.1.0